% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_rd_name.R
\name{same_rd_name}
\alias{same_rd_name}
\title{Generic test for multiple methods on the same rd file}
\usage{
same_rd_name(x)
}
\arguments{
\item{x}{A parameter}
}
\description{
Generic test for multiple methods on the same rd file
}
