% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_link.R
\name{self_link}
\alias{self_link}
\alias{self_link.default}
\title{A generic with a self-linked method}
\usage{
self_link(x)

\method{self_link}{default}(x)
}
\arguments{
\item{x}{A parameter}
}
\description{
A generic with a self-linked method
}
